Refactor for clarity; fix typing issues
