+++ bcipy/display/paradigm/vep/codes.py
@@ -1,0 +1,16 @@
+"""Functions related to VEP flash rates"""
+from typing import List
+
+import numpy as np
+
+
+def create_vep_codes(length=32, count=4) -> List[List[int]]:
+    """Create a list of random VEP codes.
+
+    length - how many bits in each code. This should be greater than or equal to the refresh rate
+        if using these to flicker. For example, if the refresh rate is 60Hz, then the length should
+        be at least 60.
+    count - how many codes to generate, each will be unique.
+    """
+    np.random.seed(1)
+    return [np.random.randint(2, size=length) for _ in range(count)]
